requirements.txt
fastapi
uvicorn
pydub
numpy
requests
websockets
python-dotenv
prometheus_client
langdetect
faster-whisper
scipy
sounddevice
celery
starlette
langchain-core
langchain-community
langchain-openai
airflow
transformers
chromadb
kafka-python==2.0.2
beautifulsoup4
lxml
aiohttp
feedparserlocust
requests